Visually illustrate the current architecture
To get my bearings I took the time to illustrate and write down the
system architecture as described by Matilda during the call on June 2nd.

My hope is by placing this within the README it makes it easy for folks
to find as the consider applying their time and attention towards the
project.